processing db interface
